requirements.txt
torch
torchvision
diffusers
transformers
accelerate
numpy
scikit-learn
Pillow
faiss-cpu